sql/postgres: move cockroach logic to its own file
